+++ internal/platform/docker.go
@@ -21,7 +21,7 @@ ENV DEBIAN_FRONTEND=noninteractive
 RUN apt-get update \
  && apt-get install -y --no-install-recommends \
 	  binutils ca-certificates coreutils cpio dosfstools dpkg e2fsprogs \
-	  file initramfs-tools kmod libarchive-tools md5deep mtools parted \
+	  dracut-core file initramfs-tools kmod libarchive-tools md5deep mtools parted \
 	  squashfs-tools xorriso xz-utils zstd \
  && rm -rf /var/lib/apt/lists/*
 `